+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Inputs can be validated using operations
   - `validate` parameter in `Input` takes `Operation.instance_name`
 - Logistic Regression with SAG optimizer
+- shouldi got an operation to run cargo-audit on rust code.
+- Moved all the downloads to tests/downloads to speed the CI test.
 - Test tensorflow DNNEstimator documentation exaples in CI
 - Add python code for tensorflow DNNEstimator
 - Ability to run a subflow as if it were an operation using the
@@ -29,7 +31,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Shouldi got an operation to run npm-audit on JavaScript code
 - Docstrings and doctestable examples for `record.py` (features and evaluated)
 - Simplified model API with SimpleModel
-- Shouldi got an operation to run cargo-aduit on RUST code.
 - Documentation on how DataFlows work conceptually.
 - Style guide now contains information on class, variable, and function naming.
 ### Changed